+++ NEWS.md
@@ -3,6 +3,7 @@
 - Use mapbayest() instead of mbrest()
 - Use get_data() instead of see_data()
 - Use get_param() to access a posteriori captured parameters
+- Use get_eta() to access eta values
 
 # mapbayr 0.3